# be-nest/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/media/{id}": {
      "get": {
        "operationId": "MediaController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      },
      "patch": {
        "operationId": "MediaController_rename",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/api/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaController_deleteMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{id}/download": {
      "get": {
        "operationId": "MediaController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/api/media/{id}/content": {
      "get": {
        "operationId": "MediaController_content",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/api/Hall/CreateHall": {
      "post": {
        "operationId": "HallController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHallDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "hall"
        ]
      }
    },
...
    "/api/Hall/UpdateHall/{hallId}": {
      "put": {
        "operationId": "HallController_update",
        "parameters": [
          {
            "name": "hallId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "hall"
        ]
      }
    },
    "/api/Hall/DeleteHall/{hallId}": {
      "delete": {
        "operationId": "HallController_delete",
        "parameters": [
          {
            "name": "hallId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "hall"
        ]
      }
    },
...
    "/api/Checkin/UpdateCheckin": {
      "put": {
        "operationId": "CheckInController_checkIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInBachelorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "check-in"
        ]
      }
    },
    "/api/Checkin/UpdateCheckinStudentCode": {
      "put": {
        "operationId": "CheckInController_checkInByStudentCode",
        "parameters": [
          {
            "name": "studentCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "check-in"
        ]
      }
    },
...
    "/api/Checkin/UpdateStatusCheckin": {
      "put": {
        "operationId": "CheckInController_updateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCheckInStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "check-in"
        ]
      }
    },
...
    "/api/Checkin/CreateCheckin": {
      "post": {
        "operationId": "CheckInController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "check-in"
        ]
      }
    },
...
    "/api/Checkin/by-hall/{hallId}": {
      "get": {
        "operationId": "CheckInController_byHall",
        "parameters": [
          {
            "name": "hallId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "check-in"
        ]
      }
    },
    "/api/Checkin/by-hall-name": {
      "get": {
        "operationId": "CheckInController_byHallName",
        "parameters": [
          {
            "name": "hallName",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "check-in"
        ]
      }
    },
...
    "/api/Mc/GetLocationBachelor": {
      "get": {
        "operationId": "LedController_location",
        "parameters": [
          {
            "name": "studentCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "led"
        ]
      }
    },
...
  },
  "info": {
    "title": "Convocation Day API",
    "description": "NestJS compatibility API for the Convocation Day frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "RenameMediaDto": {
        "type": "object",
        "properties": {}
      },
      "DeleteMediaDto": {
        "type": "object",
        "properties": {}
      },
      "CreateHallDto": {
        "type": "object",
        "properties": {}
      },
...
      "CheckInBachelorDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCheckInStatusDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCheckInDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}